renamed addInternship and added form modal
